fix: after review transfer Node class _ optimaize function InMemoryHistoryManager
